WIP rename to blast
