+++ packages/blast-cmd/init/init.js
@@ -2,8 +2,8 @@ const fs = require('fs')
 const fsExtra = require('fs-extra')
 const path = require('path')
 
-const { getPackageRootPath } = require('../../cudos-utilities/package-info')
-const CudosError = require('../../cudos-utilities/cudos-error')
+const { getPackageRootPath } = require('../../blast-utilities/package-info')
+const BlastError = require('../../blast-utilities/blast-error')
 
 async function initCmd(argv) {
   handleCustomDirCreation(argv)
@@ -13,7 +13,7 @@ async function initCmd(argv) {
       argv.dir
     )
   } catch (error) {
-    throw new CudosError(`Error copying folder: ${error}`)
+    throw new BlastError(`Error copying folder: ${error}`)
   }
 
   console.log(`Success! Sample project initialized in ${argv.dir !== '.' ? argv.dir : process.cwd()}`)
@@ -25,7 +25,7 @@ function handleCustomDirCreation(argv) {
     try {
       fs.mkdirSync(argv.dir, { recursive: true })
     } catch (error) {
-      throw new CudosError(`Error creating new directory: ${error}`)
+      throw new BlastError(`Error creating new directory: ${error}`)
     }
   }
 }
