+++ packages/blast-cmd/run/run.js
@@ -5,15 +5,15 @@ const path = require('path')
 const {
   getContractFactory,
   getContractFromAddress
-} = require('../../cudos-utilities/contract-utils')
-const CudosError = require('../../cudos-utilities/cudos-error')
+} = require('../../blast-utilities/contract-utils')
+const BlastError = require('../../blast-utilities/blast-error')
 
 global.getContractFactory = getContractFactory
 global.getContractFromAddress = getContractFromAddress
 
 async function runCmd(argv) {
   if (!fs.existsSync(`${path.resolve('.')}/${argv.scriptFilePath}`)) {
-    throw new CudosError(`Script at location ${path.resolve('.')}/${argv.scriptFilePath} does not exist.`)
+    throw new BlastError(`Script at location ${path.resolve('.')}/${argv.scriptFilePath} does not exist.`)
   }
   const ds = new vm.Script(fs.readFileSync(argv.scriptFilePath))
   return ds.runInThisContext()
