+++ packages/blast-config/init-root.sh
