+++ packages/blast-utilities/config-utils.js
@@ -1,15 +1,15 @@
 const fsExtra = require('fs-extra')
 const process = require('process')
 const path = require('path')
-const CudosError = require('./cudos-error')
+const BlastError = require('./blast-error')
 
 let config = {}
 
-const configPath = path.join(process.cwd(), 'cudos.config.js')
+const configPath = path.join(process.cwd(), 'blast.config.js')
 
 function getConfig() {
   if (!fsExtra.pathExistsSync(configPath)) {
-    throw new CudosError(`Config file was not found! Make sure that cudos.config.js exists at ${configPath}`)
+    throw new BlastError(`Config file was not found! Make sure that blast.config.js exists at ${configPath}`)
   }
   config = require(configPath)
   return config
@@ -19,7 +19,7 @@ function getAccountByName(name) {
   const { config } = getConfig()
 
   if (!config.accounts[name]) {
-    throw new CudosError('Missing Account in the config file.')
+    throw new BlastError('Missing Account in the config file.')
   }
   return config.accounts[name]
 }
@@ -28,7 +28,7 @@ function getEndpoint() {
   const { config } = getConfig()
 
   if (!config.endpoint) {
-    throw new CudosError('Missing [endpoint] in the config file.')
+    throw new BlastError('Missing [endpoint] in the config file.')
   }
   return config.endpoint
 }
@@ -37,7 +37,7 @@ function getGasPrice() {
   const { config } = getConfig()
 
   if (!config.gasPrice) {
-    throw new CudosError('Missing gasPrice in the config file.')
+    throw new BlastError('Missing gasPrice in the config file.')
   }
   return config.gasPrice
 }
