+++ template/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cudos-sample-project",
+  "name": "blast-sample-project",
   "version": "1.0.0",
   "description": "Sample template to get you started",
   "main": "index.js",
